--- packages/client/src/pages/Forum/components/CreationBlock/index.ts
@@ -1,1 +1,0 @@
-export { CreationBlock } from './CreationBlock'
