+++ packages/client/src/pages/Forum/components/CreationBlockContent/CreationBlockContent.module.scss
@@ -1,10 +1,3 @@
-.creation-block {
-  height: 64px;
-  display: flex;
-  align-items: center;
-  gap: 32px;
-}
-
 .input-block {
   display: flex;
   flex-direction: column;
