+++ packages/client/src/pages/Forum/components/CreationBlockContent/index.ts
@@ -1,0 +1,1 @@
+export { CreationBlockContent } from './CreationBlockContent'
